Refactor: Implement centralized error handling
Introduces a new `catchError` utility function to centralize error handling across various commands, components, and events. This change replaces repetitive `try...catch` blocks with a single, consistent error reporting mechanism, improving code maintainability and reducing redundancy.

+++ prefixcommands/misc/say.js
@@ -25,6 +25,6 @@ module.exports = {
           }, 10000);
         });
 
-    message.delete().then(() => message.channel.send(toSay));
+    if (message.deletable) message.delete().then(() => message.channel.send(toSay));
   },
 };
