+++ prefixcommands/misc/wow.js
@@ -25,6 +25,6 @@ module.exports = {
       .setColor('Random')
       .setImage('https://media.discordapp.net/attachments/976364997066231828/1368430209845432320/images.png');
 
-    message.delete().then(() => message.channel.send({ embeds: [embed] }));
+    if (message.deletable) message.delete().then(() => message.channel.send({ embeds: [embed] }));
   },
 };
